+++ src/ric/2C4C4.c
@@ -1,0 +1,381 @@
+#include "ric.h"
+
+// Entity ID #7. Made by blueprint 6. Comes from subweapon 3. Holy water!
+// Not at all the same as DRA's.
+void EntityHolyWater(Entity* self) {
+    s16 argY;
+    s32 collision_result;
+    s32 tempXVel;
+    s32 trigresult;
+    s32 trigtemp;
+
+    if (self->step >= 3) {
+        self->posY.i.hi += 5;
+    }
+    switch (self->step) {
+    case 0:
+        self->flags = FLAG_UNK_08000000;
+        self->animSet = ANIMSET_OVL(0x12);
+        self->animCurFrame = 0x23;
+        self->zPriority = PLAYER.zPriority + 2;
+        self->unk5A = 0x46;
+        self->palette = 0x812F;
+        self->posX.i.hi = self->posX.i.hi;
+        self->posY.i.hi = self->posY.i.hi - 0x10;
+        self->ext.holywater.angle = (rand() & 0x7F) + 0xDC0;
+        if (PLAYER.facingLeft == 1) {
+            self->ext.holywater.angle = (rand() & 0x7F) + 0x9C0;
+        }
+
+        trigresult = rcos(self->ext.holywater.angle);
+        trigtemp = trigresult * 16;
+        self->velocityX = (trigresult * 32 + trigtemp) << 9 >> 8;
+
+        trigresult = rsin(self->ext.holywater.angle);
+        trigtemp = trigresult * 16;
+        self->velocityY = -((trigresult * 32 + trigtemp) << 9) >> 8;
+
+        self->ext.factory.unkB0 = 3;
+        func_8015FAB8(self);
+        self->hitboxWidth = 4;
+        self->hitboxHeight = 4;
+        self->ext.holywater.unk80 = 0x200;
+        self->step += 1;
+        break;
+    case 1:
+        self->posY.val += self->velocityY;
+        collision_result = func_801682B4(0, 0);
+        self->posX.val += self->velocityX;
+
+        if (self->velocityX < 0) {
+            argY = -4;
+        } else {
+            argY = 4;
+        }
+        collision_result |= func_8016840C(-7, argY);
+        if (collision_result & 2) {
+            self->velocityX = -self->velocityX;
+            collision_result = 1;
+        }
+        if ((collision_result & 1) || (self->hitFlags != 0)) {
+            g_api.CreateEntFactoryFromEntity(self, 0x28U, 0);
+            g_api.PlaySfx(0x68A);
+            self->ext.holywater.timer = 0x50;
+            self->animSet = 0;
+            self->hitboxState = 0;
+            self->step = 3;
+            self->velocityX >>= 2;
+        }
+        break;
+    case 2:
+        if (--self->ext.holywater.timer == 0) {
+            self->ext.holywater.timer = 0x50;
+            self->velocityX >>= 2;
+            self->step++;
+        }
+        break;
+    case 3:
+        if (!(self->ext.holywater.timer & 3)) {
+            CreateEntFactoryFromEntity(
+                self, (self->ext.holywater.unk82 << 0x10) | 7, 0);
+            // weirdly nonmatching with ++;
+            self->ext.holywater.unk82 = self->ext.holywater.unk82 + 1;
+            self->velocityX -= (self->velocityX / 32);
+        }
+        collision_result = func_801682B4(6, 0);
+        tempXVel = self->velocityX;
+        if (self->velocityX < 0) {
+            if ((collision_result & 0xF000) == 0xC000) {
+                tempXVel = tempXVel * 10 / 16;
+            }
+            if ((collision_result & 0xF000) == 0xD000) {
+                tempXVel = tempXVel * 13 / 16;
+            }
+        }
+        if (self->velocityX > 0) {
+            if ((collision_result & 0xF000) == 0x8000) {
+                tempXVel = tempXVel * 10 / 16;
+            }
+            if ((collision_result & 0xF000) == 0x9000) {
+                tempXVel = tempXVel * 13 / 16;
+            }
+        }
+        self->posX.val += tempXVel;
+        if (self->velocityX < 0) {
+            argY = -4;
+        } else {
+            argY = 4;
+        }
+        collision_result |= func_8016840C(-7, argY);
+        if (!(collision_result & 1)) {
+            self->velocityX >>= 1;
+            self->step++;
+        }
+        break;
+    case 4:
+        if (!(self->ext.holywater.timer & 3)) {
+            CreateEntFactoryFromEntity(
+                self, FACTORY(self->ext.holywater.unk82 * 0x100, 7), 0);
+            self->ext.holywater.unk82 = self->ext.holywater.unk82 + 1;
+        }
+        self->velocityY += FIX(12.0 / 128);
+        if (self->velocityY > FIX(4)) {
+            self->velocityY = FIX(4);
+        }
+        self->posY.val += self->velocityY;
+        collision_result = func_801682B4(0, 0);
+        self->posX.val += self->velocityX;
+        if (self->velocityX < 0) {
+            argY = -4;
+        } else {
+            argY = 4;
+        }
+        collision_result |= func_8016840C(-7, argY);
+        if ((collision_result & 1)) {
+            self->velocityX *= 2;
+            self->step--;
+        }
+        break;
+    case 5:
+        break;
+    }
+    if (self->step >= 3) {
+        if (--self->ext.holywater.timer < 0) {
+            DestroyEntity(self);
+            return;
+        }
+        if (self->ext.holywater.timer == 2) {
+            self->step = 5;
+        }
+        self->animCurFrame = 0;
+        self->hitboxState = 0;
+        self->posY.i.hi -= 5;
+    }
+    g_Player.D_80072F00[3] = 2;
+}
+
+INCLUDE_ASM("ric/nonmatchings/2C4C4", func_80168A20);
+
+INCLUDE_ASM("ric/nonmatchings/2C4C4", func_8016902C);
+
+INCLUDE_ASM("ric/nonmatchings/2C4C4", func_80169470);
+
+// ID 3. Created by blueprint 2.
+INCLUDE_ASM("ric/nonmatchings/2C4C4", func_80169704);
+
+void func_80169C10(Entity* entity) {
+    Primitive* prim;
+    s16 primIndex;
+    s32 PosX = 8;
+    s32 PosY = 4;
+
+    switch (entity->step) {
+    case 0:
+        primIndex = g_api.AllocPrimitives(PRIM_GT4, 1);
+        entity->primIndex = primIndex;
+        if (primIndex != -1) {
+            entity->flags = FLAG_UNK_08000000 | FLAG_HAS_PRIMS;
+            entity->velocityY = FIX(0.5);
+            entity->posX.i.hi =
+                ((u16)entity->posX.i.hi - PosX) + (rand() & 0xF);
+            entity->posY.i.hi =
+                ((u16)entity->posY.i.hi - PosY) + (rand() & 0xF);
+            prim = &g_PrimBuf[entity->primIndex];
+            prim->clut = 0x1B0;
+            prim->tpage = 0x1A;
+            prim->b0 = 0;
+            prim->b1 = 0;
+            prim->priority = entity->zPriority;
+            prim->priority = prim->priority + 4;
+            prim->drawMode = 0x31;
+            func_8015FDB0(prim, entity->posX.i.hi, entity->posY.i.hi);
+            entity->step++;
+        } else {
+            DestroyEntity(entity);
+        }
+        break;
+
+    default:
+        entity->posY.val += entity->velocityY;
+        prim = &g_PrimBuf[entity->primIndex];
+        if (func_8015FDB0(prim, entity->posX.i.hi, entity->posY.i.hi) != 0) {
+            DestroyEntity(entity);
+        }
+        break;
+    }
+}
+
+void func_80169D74(Entity* entity) {
+    Multi temp;
+    s16* ptr;
+
+    switch (entity->step) {
+    case 0:
+        entity->flags = FLAG_UNK_04000000 | FLAG_UNK_08000000;
+        entity->ext.generic.unk84.unk =
+            entity->ext.generic.unk8C.entityPtr->ext.generic.unk84.unk;
+        entity->animSet = ANIMSET_OVL(17);
+        entity->animCurFrame = D_80155E68[entity->params];
+        entity->unk5A = 0x66;
+        entity->palette = 0x81B0;
+        entity->drawMode = DRAW_TPAGE;
+        entity->facingLeft = PLAYER.facingLeft;
+        entity->zPriority = PLAYER.zPriority;
+        entity->drawFlags = FLAG_DRAW_ROTZ;
+        entity->rotZ = 0xC00;
+        entity->step++;
+        break;
+
+    case 1:
+        entity->rotZ -= 0x80;
+        if (entity->ext.generic.unk8C.entityPtr->step == 7) {
+            entity->step++;
+            entity->ext.generic.unk7C.s = (entity->params + 1) * 4;
+        }
+        break;
+
+    case 2:
+        entity->rotZ -= 0x80;
+        entity->ext.generic.unk7C.s--;
+        if (entity->ext.generic.unk7C.s == 0) {
+            DestroyEntity(entity);
+            return;
+        }
+        break;
+    }
+    temp = entity->ext.generic.unk84;
+    ptr = temp.unk + ((u16)entity->ext.generic.unk80.modeS16.unk0 * 4);
+    entity->posX.i.hi = ptr[0] - g_Tilemap.scrollX.i.hi;
+    entity->posY.i.hi = ptr[1] - g_Tilemap.scrollY.i.hi;
+    entity->ext.generic.unk80.modeS16.unk0 =
+        (entity->ext.generic.unk80.modeS16.unk0 + 1) & 0x3F;
+}
+
+INCLUDE_ASM("ric/nonmatchings/2C4C4", func_80169F04);
+
+INCLUDE_ASM("ric/nonmatchings/2C4C4", func_8016A26C);
+
+INCLUDE_ASM("ric/nonmatchings/2C4C4", func_8016A974);
+
+INCLUDE_ASM("ric/nonmatchings/2C4C4", func_8016B0C0);
+
+void func_8016B8E8(s32 arg0) {
+    g_CurrentEntity->ext.generic.unk7C.s =
+        (arg0 << 0x10 >> 0xF) - g_CurrentEntity->ext.generic.unk7C.s;
+    if (g_CurrentEntity->ext.generic.unk80.modeS16.unk2 == 0) {
+        g_CurrentEntity->ext.generic.unk80.modeS16.unk0++;
+        g_CurrentEntity->ext.generic.unk80.modeS16.unk2++;
+    }
+}
+
+void func_8016B92C(s16 arg0) {
+    if (g_CurrentEntity->ext.generic.unk80.modeS16.unk2 == 0) {
+        g_CurrentEntity->ext.generic.unk7C.s =
+            (arg0 * 2) - g_CurrentEntity->ext.generic.unk7C.s;
+        g_CurrentEntity->ext.generic.unk80.modeS16.unk0++;
+        g_CurrentEntity->ext.generic.unk80.modeS16.unk2++;
+    }
+}
+
+INCLUDE_ASM("ric/nonmatchings/2C4C4", func_8016B97C);
+
+INCLUDE_ASM("ric/nonmatchings/2C4C4", func_8016C1BC);
+
+s32 RicPrimDecreaseBrightness(Primitive* prim, u8 amount) {
+    u8 isEnd;
+    s32 i;
+    s32 j;
+    u8* pColor;
+
+    isEnd = 0;
+    pColor = &prim->r0;
+    for (i = 0; i < 4; i++, pColor += OFF(Primitive, r1) - OFF(Primitive, r0)) {
+        for (j = 0; j < 3; j++) {
+            pColor[j] -= amount;
+
+            if (pColor[j] < 16) {
+                pColor[j] = 16;
+            } else {
+                isEnd |= 1;
+            }
+        }
+    }
+    return isEnd;
+}
+
+INCLUDE_ASM("ric/nonmatchings/2C4C4", func_8016C734);
+
+INCLUDE_ASM("ric/nonmatchings/2C4C4", func_8016CC74);
+
+void func_8016D328(Entity* entity) {
+    s16 primIndex;
+    s32 newVelocity;
+
+    switch (entity->step) {
+    case 0:
+        primIndex = g_api.AllocPrimitives(PRIM_GT4, 1);
+        entity->primIndex = primIndex;
+        if (primIndex != -1) {
+            entity->flags = FLAG_UNK_08000000 | FLAG_HAS_PRIMS;
+            entity->posX.val =
+                entity->ext.generic.unk8C.entityPtr->ext.generic.unk84.unk;
+            entity->posY.val =
+                entity->ext.generic.unk8C.entityPtr->ext.generic.unk88.unk;
+            entity->facingLeft = entity->ext.generic.unk8C.entityPtr->ext
+                                     .generic.unk8C.modeU16.unk0;
+            entity->ext.generic.unkB0 = 0x18;
+            func_8015FAB8(entity);
+            entity->unk5A = 0x79;
+            entity->animSet = ANIMSET_DRA(14);
+            entity->palette = 0x819E;
+            entity->unk4C = &D_80155EA8;
+            entity->drawMode = 0x30;
+            entity->drawFlags = FLAG_DRAW_UNK8;
+            entity->unk6C = 0x60;
+            entity->hitboxWidth = 8;
+            entity->hitboxHeight = 8;
+            entity->flags |= FLAG_UNK_100000;
+            newVelocity = (rand() % 512) + 0x300;
+            entity->velocityX = rcos(newVelocity) * 32;
+            entity->velocityY = -(rsin(newVelocity) * 32);
+            entity->step++;
+        } else {
+            DestroyEntity(entity);
+        }
+        break;
+
+    case 1:
+        if (++entity->ext.generic.unk7C.s >= 39) {
+            DestroyEntity(entity);
+        } else {
+            entity->posX.val += entity->velocityX;
+            entity->posY.val += entity->velocityY;
+        }
+        break;
+    }
+}
+
+INCLUDE_ASM("ric/nonmatchings/2C4C4", func_8016D4D8);
+
+void func_8016D920(Entity* entity) {
+    switch (entity->step) {
+    case 0:
+        entity->flags = FLAG_UNK_04000000;
+        entity->ext.generic.unkB0 = 0x19;
+        func_8015FAB8(entity);
+        entity->hitboxWidth = 4;
+        entity->hitboxHeight = 4;
+        entity->step++;
+        break;
+
+    case 1:
+        if (++entity->ext.generic.unk7C.s >= 4) {
+            DestroyEntity(entity);
+        }
+        break;
+    }
+}
+
+INCLUDE_ASM("ric/nonmatchings/2C4C4", func_8016D9C4);
+
+INCLUDE_ASM("ric/nonmatchings/2C4C4", func_8016DF74);
